move signup logic to user service
